Starlight.BackgroundDaemon/pyproject.toml
[project]
name = "brainbit-background-daemon"
version = "7.2.7"
description = "Yes."
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pyneurosdk2==1.0.14", "websockets==14.2"]
